rename tests/data to tests/testdata
