+++ tests/testdata/trainercontroller/epoch-level-training-loss.yaml
