+++ tests/testdata/trainercontroller/exposed_metrics.yaml
