+++ tests/testdata/trainercontroller/incorrect_source_event_exposed_metrics.yaml
