+++ tests/testdata/trainercontroller/log_controller.yaml
