+++ tests/testdata/trainercontroller/loss_custom_operation.yaml
