+++ tests/testdata/trainercontroller/loss_invalid_metric.yaml
