+++ tests/testdata/trainercontroller/loss_invalid_operation.yaml
