+++ tests/testdata/trainercontroller/loss_invalid_trigger.yaml
