+++ tests/testdata/trainercontroller/loss_on_threshold.yaml
