+++ tests/testdata/trainercontroller/loss_unavailable_metric.yaml
