+++ tests/testdata/trainercontroller/loss_with_invalid_type_rule.yaml
