+++ tests/testdata/trainercontroller/loss_with_malicious_input_rule.yaml
