+++ tests/testdata/trainercontroller/loss_with_malicious_os_rule.yaml
