+++ tests/testdata/trainercontroller/thresholded-training-loss.yaml
